+++ apps/cli/package.json
@@ -21,7 +21,7 @@
   "dependencies": {
     "@trpc/client": "^11.8.1",
     "commander": "^13.1.0",
-    "diff": "^7.0.0",
+    "diff": "^8.0.3",
     "fast-glob": "^3.3.3",
     "picocolors": "^1.1.1",
     "superjson": "^2.2.6",
@@ -29,7 +29,7 @@
   },
   "devDependencies": {
     "@lobechat/device-gateway-client": "workspace:*",
-    "@types/diff": "^6.0.0",
+    "@lobechat/local-file-shell": "workspace:*",
     "@types/node": "^22.13.5",
     "@types/ws": "^8.18.1",
     "tsup": "^8.4.0",
