+++ apps/cli/tsup.config.ts
@@ -4,8 +4,15 @@ export default defineConfig({
   banner: { js: '#!/usr/bin/env node' },
   clean: true,
   entry: ['src/index.ts'],
+  external: ['@napi-rs/canvas', 'fast-glob', 'diff', 'debug'],
   format: ['esm'],
-  noExternal: ['@lobechat/device-gateway-client', '@trpc/client', 'superjson'],
+  noExternal: [
+    '@lobechat/device-gateway-client',
+    '@lobechat/local-file-shell',
+    '@lobechat/file-loaders',
+    '@trpc/client',
+    'superjson',
+  ],
   platform: 'node',
   target: 'node18',
 });
