+++ packages/local-file-shell/src/file/index.ts
@@ -1,0 +1,8 @@
+export { editLocalFile } from './edit';
+export { globLocalFiles } from './glob';
+export { grepContent } from './grep';
+export type { ListFilesOptions } from './list';
+export { listLocalFiles } from './list';
+export { readLocalFile } from './read';
+export { searchLocalFiles } from './search';
+export { writeLocalFile } from './write';
