+++ packages/local-file-shell/src/index.ts
@@ -1,0 +1,3 @@
+export * from './file';
+export * from './shell';
+export * from './types';
