+++ packages/local-file-shell/src/shell/index.ts
@@ -1,0 +1,5 @@
+export type { ShellProcess } from './process-manager';
+export { ShellProcessManager } from './process-manager';
+export type { RunCommandOptions } from './runner';
+export { runCommand } from './runner';
+export { getShellConfig, MAX_OUTPUT_LENGTH, stripAnsi, truncateOutput } from './utils';
